Fixing some mistakes.
