--- dz2/diff_test.sh
@@ -1,12 +1,0 @@
-./../build/st_main.out ../tests a > stat.txt
-./../build/dyn_main.out ../tests a > dyn.txt
-cat stat.txt
-cat dyn.txt
-./../build/st_main.out ../tests You > stat.txt
-./../build/dyn_main.out ../tests You > dyn.txt
-cat stat.txt
-cat dyn.txt
-./../build/st_main.out ../tests _ancbschsd > stat.txt
-./../build/dyn_main.out ../tests _ancbschsd > dyn.txt
-cat stat.txt
-cat dyn.txt?
