+++ dz2/extra_func.h
@@ -3,6 +3,9 @@
 
 #include <stdio.h>
 #include <string.h>
+#include <stdlib.h>
+#include <dirent.h>
+#include <unistd.h>
 
 #define OK 0
 #define ERR_IO -1
@@ -10,12 +13,25 @@
 #define ERR_PARAM -3
 #define ERR_ARG -4
 #define TOP_COUNT 5
+#define PATH_MAX 4096
 
 typedef struct files_t
 {
     int count[TOP_COUNT];
     char *name[TOP_COUNT];
 } topfiles;
+
+typedef struct array_t
+{
+    char **data;
+    int len;
+    int size;
+    int step;
+} dyn_array;
+
+int get_filenames_array(dyn_array *f_names, const char *dir_name);
+
+int free_filenames_array(dyn_array *f_names);
 
 int get_str_count(const char *file, const char *str, int *count);
